--- app/src/main/res/layout/activity_main.xml
@@ -1,117 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context="nz.org.cacophony.sidekick.MainActivity">
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:orientation="vertical">
-
-        <LinearLayout
-            android:id="@+id/network_error_message_layout"
-            android:visibility="gone"
-            android:layout_width="match_parent"
-            android:background="@color/alert"
-            android:layout_height="70dp">
-
-            <TextView
-                android:id="@+id/network_error_message_text"
-                android:padding="8dp"
-                android:layout_weight="1"
-                android:layout_width="0dp"
-                android:layout_height="wrap_content"
-                android:layout_gravity="center"
-                android:textSize="18sp"
-                android:textStyle="bold" />
-
-            <Button
-                android:layout_gravity="center"
-                android:onClick="enableValidAp"
-                android:layout_width="130dp"
-                android:textAlignment="center"
-                android:text="enable hotspot"
-                android:layout_height="wrap_content" />
-        </LinearLayout>
-
-        <LinearLayout
-            android:id="@+id/network_warning_message_layout"
-            android:visibility="gone"
-            android:layout_width="match_parent"
-            android:background="@color/warning"
-            android:layout_height="70dp">
-
-            <TextView
-                android:id="@+id/network_warning_message_text"
-                android:padding="8dp"
-                android:text=""
-                android:layout_weight="1"
-                android:layout_width="0dp"
-                android:layout_height="wrap_content"
-                android:layout_gravity="center"
-                android:textSize="18sp"
-                android:textStyle="bold" />
-
-            <Button
-                android:layout_gravity="center"
-                android:onClick="openNetworkSettings"
-                android:layout_width="130dp"
-                android:textAlignment="center"
-                android:text="Network settings"
-                android:layout_height="wrap_content" />
-        </LinearLayout>
-
-        <ProgressBar
-            android:id="@+id/progressBar"
-            style="@style/Widget.AppCompat.ProgressBar.Horizontal"
-            android:layout_width="match_parent"
-            android:layout_height="15dp"
-            android:indeterminate="true"
-            android:visibility="visible"
-            tools:layout_editor_absoluteX="8dp" />
-
-        <TextView
-            android:id="@+id/placeholder_text"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="Searching for devices..."
-            android:padding="8dp"
-            android:textSize="24sp" />
-
-        <androidx.recyclerview.widget.RecyclerView
-            android:id="@+id/device_list"
-            android:layout_width="match_parent"
-            android:layout_height="0dp"
-            android:padding="8dp"
-            android:scrollbars="vertical"
-            android:layout_weight="1"
-            tools:layout_editor_absoluteY="47dp" />
-
-        <Button
-            android:id="@+id/upload_recordings_button"
-            android:onClick="uploadRecordings"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:text="Upload recordings"/>
-
-        <Button
-            android:layout_marginTop="30dp"
-            android:id="@+id/set_location"
-            android:onClick="setLocationButton"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:text="Update All Devices Location"/>
-
-        <Button
-            android:layout_marginTop="30dp"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:onClick="refreshDevices"
-            android:text="Refresh" />
-
-    </LinearLayout>
-
-</androidx.constraintlayout.widget.ConstraintLayout>
