--- app/src/main/res/menu/settings.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<menu xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
-<item
-    android:id="@+id/settings_button"
-    android:title="Settings"
-    app:showAsAction="always"
-    android:icon="@drawable/ic_settings_black_24dp"
-    />
-</menu>
